(MAINT) Rework site to include blog
This commit is very difficult to parse as it reorganizes the
jekyll site to have a front-page blog and a linked page for
the tools. Any new blog posts will automatically publish to
the front page, any new tools added to the yaml data file
will be added to the list on the tools page.

This commit also removes erroneous files not needed for the
github pages to work.