Update tools to have a separate page each
With this change in place, we can have a full page explaining (necessarily duplicating) the reason for using a tool. This gives us a bit more space and flexibility with the tool's page.

The tools.json output could be used as input for community_management scripts (and thereby making this the canonical source), but would require more work than just `jsonify` to work.